+++ CHANGES.rst
@@ -8,6 +8,8 @@
   compatible with Python 3.6 or later and ``astropy`` 3.x or later. [#243]
 - Added support for ``RickerWavelet1D`` model that is the renamed version
   of ``MexicanHat1D`` model to be consistent with ``astropy`` 4.0. [#250]
+- Added option for ``synphot.utils.download_data()`` to download to the cache
+  instead of a specific location. [#211]
 
 0.2.1 (2019-12-20)
 ==================